# docs/swaggerOpenapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Toiage Core — Educational Intelligence Platform",
    "description": "API for generating curriculum-aligned science projects, assessments, and educational content for CBSE classes 6-8.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/teacher-assistant/generate": {
      "post": {
        "tags": [
          "teacher-assistant"
        ],
        "summary": "Generate assessment with teacher-specified pattern",
        "description": "Generate a CBSE-aligned assessment where the teacher specifies question types, counts, and marks per question. The pattern is auto-saved for reuse. Supports any question types: mcq, short_answer, long_answer, fill_blanks, true_false, match, etc.",
        "operationId": "generate_assessment_api_v1_teacher_assistant_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateAssessmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateAssessmentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/teacher-assistant/history/{history_id}": {
      "get": {
        "tags": [
          "teacher-assistant"
        ],
        "summary": "Get generation by ID",
        "description": "Retrieve a specific generation with full section details.",
        "operationId": "get_generation_api_v1_teacher_assistant_history__history_id__get",
        "parameters": [
          {
            "name": "history_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "History Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateAssessmentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Generation not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExamSection": {
        "properties": {
          "section_name": {
            "type": "string",
            "title": "Section Name",
            "description": "Section name"
          },
          "marks_per_question": {
            "type": "integer",
            "title": "Marks Per Question",
            "description": "Marks per question"
          },
          "total_section_marks": {
            "type": "integer",
            "title": "Total Section Marks",
            "description": "Total marks for this section"
          },
          "instructions": {
            "type": "string",
            "title": "Instructions",
            "description": "Section instructions"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "type": "array",
            "title": "Questions",
            "description": "Questions in this section"
          }
        },
        "type": "object",
        "required": [
          "section_name",
          "marks_per_question",
          "total_section_marks",
          "instructions"
        ],
        "title": "ExamSection",
        "description": "A section of an exam paper."
      },
      "GenerateAssessmentRequest": {
        "properties": {
          "grade": {
            "type": "integer",
            "title": "Grade",
            "description": "CBSE class grade",
            "examples": [6]
          },
          "subject": {
            "type": "string",
            "title": "Subject",
            "description": "Subject name",
            "examples": [
              "Science"
            ]
          },
          "chapter": {
            "type": "string",
            "title": "Chapter",
            "description": "Chapter name",
            "examples": [
              "Water"
            ]
          },
          "topic": {
            "type": "string",
            "title": "Topic",
            "description": "Topic",
            "default": "",
            "examples": [
              "Water Conservation"
            ]
          },
          "question_specs": {
            "items": {
              "$ref": "#/components/schemas/QuestionSpecItem"
            },
            "type": "array",
            "title": "Question Specs",
            "description": "List of question type specifications"
          },
          "pattern_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pattern Name",
            "description": "Optional name to save this pattern as. Auto-generated if not provided.",
            "examples": [
              "Class 6 Science Annual"
            ]
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes",
            "description": "Recommended duration in minutes"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty",
            "default": "medium"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "mock"
          }
        },
        "type": "object",
        "required": [
          "grade",
          "subject",
          "chapter",
          "question_specs"
        ],
        "title": "GenerateAssessmentRequest",
        "description": "Request to generate a custom assessment — replaces all previous specific endpoints."
      },
      "GenerateAssessmentResponse": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject",
            "description": "Subject name"
          },
          "grade": {
            "type": "integer",
            "title": "Grade",
            "description": "CBSE class grade"
          },
          "chapter": {
            "type": "string",
            "title": "Chapter",
            "description": "Chapter name"
          },
          "topic": {
            "type": "string",
            "title": "Topic",
            "description": "Topic"
          },
          "total_marks": {
            "type": "integer",
            "title": "Total Marks",
            "description": "Total marks"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/ExamSection"
            },
            "type": "array",
            "title": "Sections",
            "description": "Assessment sections"
          },
          "total_time_minutes": {
            "type": "string",
            "title": "Total Time Minutes",
            "description": "Recommended time"
          },
          "instructions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Instructions",
            "description": "General instructions"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "Generation history ID",
            "examples": [42]
          },
          "pattern_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pattern Id",
            "description": "Saved pattern ID",
            "examples": [5]
          },
          "pattern_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pattern Name",
            "description": "Name of the saved pattern"
          }
        },
        "type": "object",
        "required": [
          "subject",
          "grade",
          "chapter",
          "topic",
          "total_marks",
          "total_time_minutes"
        ],
        "title": "GenerateAssessmentResponse",
        "description": "Response for assessment generation — includes IDs for pattern and generation history."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MCQOption": {
        "properties": {
          "A": {
            "type": "string",
            "title": "A",
            "description": "Option A"
          },
          "B": {
            "type": "string",
            "title": "B",
            "description": "Option B"
          },
          "C": {
            "type": "string",
            "title": "C",
            "description": "Option C"
          },
          "D": {
            "type": "string",
            "title": "D",
            "description": "Option D"
          }
        },
        "type": "object",
        "required": [
          "A",
          "B",
          "C",
          "D"
        ],
        "title": "MCQOption",
        "description": "Options for a multiple-choice question."
      },
...
      "Question": {
        "properties": {
          "question_number": {
            "type": "integer",
            "title": "Question Number",
            "description": "Question number"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text",
            "description": "Question text"
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MCQOption"
              },
              {
                "type": "null"
              }
            ],
            "description": "Options for MCQ"
          },
          "correct_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correct Answer",
            "description": "Correct answer"
          },
          "marks": {
            "type": "integer",
            "title": "Marks",
            "description": "Marks for this question",
            "default": 1
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty",
            "description": "Difficulty level",
            "default": "medium"
          },
          "cognitive_level": {
            "type": "string",
            "title": "Cognitive Level",
            "description": "Cognitive level",
            "default": "recall"
          },
          "chapter": {
            "type": "string",
            "title": "Chapter",
            "description": "Chapter name",
            "default": ""
          },
          "model_explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Explanation",
            "description": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "question_number",
          "question_text"
        ],
        "title": "Question",
        "description": "A generic question (for exam papers)."
      },
      "QuestionSpecItem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Question type: mcq, short_answer, long_answer, fill_blanks, true_false, match, etc.",
            "examples": [
              "mcq",
              "short_answer",
              "long_answer"
            ]
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Number of questions of this type",
            "examples": [5]
          },
          "marks_per_question": {
            "type": "integer",
            "title": "Marks Per Question",
            "description": "Marks per question",
            "examples": [1, 2, 3, 5]
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty",
            "description": "Difficulty for this section",
            "default": "medium"
          },
          "cognitive_levels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cognitive Levels",
            "description": "Cognitive levels to cover"
          }
        },
        "type": "object",
        "required": [
          "type",
          "count",
          "marks_per_question"
        ],
        "title": "QuestionSpecItem",
        "description": "Specification for a single question type in an assessment."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}